refactor(gemini): align client manager ownership with provider package
- move Gemini client manager implementation to providers/gemini/client.py
- replace services/gemini_client.py with a compatibility shim
- update Gemini provider components to use direct intra-package imports
- update application startup wiring to use the new authoritative client path
- preserve backward compatibility for legacy endpoints and existing test mocks
- maintain existing initialization, authentication, and recovery behavior
- verify no circular imports introduced by relocation

Architecture:
- completes Gemini ownership alignment after AuthManager separation
- encapsulates Gemini-specific client lifecycle within the Gemini provider package
- keeps compatibility layer in place for gradual migration

Tested:
- full test suite passing (127 passed)
