+++ src/app/services/providers/gemini/session_manager.py
@@ -7,7 +7,7 @@ from datetime import datetime, timedelta, timezone
 from typing import Dict, Optional, Any, AsyncGenerator, List
 from app.config import get_default_conversation_snapshot_db
 from app.logger import logger
-from app.services.gemini_client import get_gemini_client, GeminiClientNotInitializedError
+from app.services.providers.gemini.client import get_gemini_client, GeminiClientNotInitializedError
 from app.services.providers.base_repository import ConversationSnapshot, IConversationRepository, ProviderCapability
 from app.services.providers.exceptions import SnapshotNotFoundError, StateIntegrityError
 from app.utils.tokens import generate_opaque_token
