+++ src/app/services/providers/gemini/webapi_adapter.py
@@ -5,7 +5,7 @@ from fastapi import HTTPException
 from fastapi.responses import StreamingResponse
 from gemini_webapi.exceptions import APIError
 
-from app.services.gemini_client import get_gemini_client, GeminiClientNotInitializedError
+from app.services.providers.gemini.client import get_gemini_client, GeminiClientNotInitializedError
 from app.services.providers.gemini.session_manager import get_gemini_chat_registry
 from app.services.providers.exceptions import SessionRecoveryError, SnapshotNotFoundError
 from app.services.providers.gemini.base_adapter import GeminiBackendAdapter
